requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
decorator==5.1.1
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jupyter
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
torchaudio==2.5.0
triton==3.1.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
